+++ src/app/js/app.js
@@ -19,6 +19,7 @@
     ,'orrportal.st'
     ,'orrportal.kw'
     ,'orrportal.v2r'
+    ,'orrportal.m2r'
     ,'orrportal.voc'
     ,'orrportal.firebase'
     ,'orrportal.upload'